LayoutTests/storage/indexeddb/setversion-blocks-setversion-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test that two setVersions from the same connection execute sequentially.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "setversion-blocks-setversion.html"
ver = 1
versionChangeComplete = false
indexedDB.open(dbname)

h1OpenOnSuccess():
h1 = event.target.result
h1.setVersion(String(ver++))
h1.setVersion(String(ver++))

h1SetVersionOnSuccess1():

transactionOnComplete1():
half done
versionChangeComplete = true

h1SetVersionOnSuccess2():

transactionOnComplete2():
PASS versionChangeComplete is true
PASS successfullyParsed is true

TEST COMPLETE
